wmaker: add core workspace pager functions
This patch is adding the core code needed to run the workspace pager.

THIS PATCH WAS MODIFIED BY Rodolfo García Peñas (kix) because cannot be
applied in the awmaker git without changes. This patch is based in the
original patch created by David Maciejak <david.maciejak@gmail.com>.

This patch has a lot of changes, first the support for vscr, and changes
about code style for the new files wsmap.[ch].