wmaker: removed unused constant SCROLL_STEPS in the switchpanel code
As pointed by gcc, this constant is not used anywhere in the code. It seems
to have been prepared for a scroll animation when the switchpanel would be
too big to fit the screen, but the code looks like to never have been
activated, and was cleaned away in commit
  6d08aa22d824d6abe5edb4c4892bbf9eaec3175b

If the feature were to be re-added, it would be better to use something
proportional to the menu scroll speed parameter that can be configured by
user anyway.

Signed-off-by: Christophe CURIS <christophe.curis@free.fr>