VSCR: WScreen and virtual screens are independent
This patch is the final (only a few hacks more are needed, but is
the most important) patch to make the WScreen struct and the virtual
screen struct independent.

Now, when awmaker is launched, it read the WScreens and create them.
Then, only using the number of WScreens, the virtual screens are
created.

In future patches, we can create or not the same number of virtual
screens or use the previously created virtual screens (p.e. restart).

Now, the virtual screen do not depends on WScreen. Now we can create
the virtual screen and then, the WScreens.