i18n: fix failing generation of the README.i18n file
The command in 'autogen.sh' was using an invalid syntax to call the script,
so the file was not generated; this patch fixes the invocation.

The test in the makefile was not correct because it would not generate the
file if it was deleted; now the check works whether the file exists or not,
and properly skips generation if it is read-only (for make distcheck).

Signed-off-by: Christophe CURIS <christophe.curis@free.fr>